Added sql to create updated database tables and added a script to automatically create the tables and drop them if they already exist

+++ database/procedures/001_create_user.sql
@@ -9,7 +9,7 @@ BEGIN
   SET NOCOUNT ON;
   SET XACT_ABORT ON;
 
-  DECLARE @UserID INT;
+  DECLARE @UserID BIGINT;
 
   SET @Username = LTRIM(RTRIM(@Username));
   SET @PasswordHash = LTRIM(RTRIM(@PasswordHash));
